hw10-view v3

+++ hw08-view/src/main/java/ru/cr/hw/services/BookServiceImpl.java
@@ -4,13 +4,15 @@ import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import ru.cr.hw.controller.NotFoundException;
 import ru.cr.hw.domain.Book;
+import ru.cr.hw.dto.BookDto;
 import ru.cr.hw.repostory.AuthorRepository;
 import ru.cr.hw.repostory.BookRepository;
 import ru.cr.hw.repostory.GenreRepository;
 
 import java.util.List;
-import java.util.Optional;
+import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
 @Service
@@ -22,29 +24,33 @@ public class BookServiceImpl implements BookService {
     private final BookRepository bookRepository;
 
     @Override
-    public Optional<Book> findById(long id) {
-        return bookRepository.findById(id);
+    public BookDto findById(long id) {
+        Book book = bookRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Book with id %d not found".formatted(id)));
+        return BookDto.fromDomain(book);
     }
 
     @Override
-    public List<Book> findAll() {
-        return bookRepository.findAll();
+    public List<BookDto> findAll() {
+        return bookRepository.findAll().stream()
+                .map(BookDto::fromDomain)
+                .collect(Collectors.toList());
     }
 
     @Override
-    public Book insert(String title, long authorId, long genreId) {
+    public BookDto insert(String title, long authorId, long genreId) {
         var author = authorRepository.findById(authorId)
                 .orElseThrow(() -> new EntityNotFoundException("Author with id %d not found".formatted(authorId)));
         var genre = genreRepository.findById(genreId)
                 .orElseThrow(() -> new EntityNotFoundException("Genre with id %d not found".formatted(genreId)));
-        Book book;
-        book = new Book(title, author, genre);
-        return bookRepository.save(book);
+        Book book = new Book(title, author, genre);
+        Book saved = bookRepository.save(book);
+        return BookDto.fromDomain(saved);
     }
 
     @Override
     @Transactional
-    public Book update(long id, String title, long authorId, long genreId) {
+    public BookDto update(long id, String title, long authorId, long genreId) {
         var author = authorRepository.findById(authorId)
                 .orElseThrow(() -> new EntityNotFoundException("Author with id %d not found".formatted(authorId)));
         var genre = genreRepository.findById(genreId)
@@ -56,11 +62,16 @@ public class BookServiceImpl implements BookService {
         book.setAuthor(author);
         book.setGenre(genre);
 
-        return bookRepository.save(book);
+        Book updated = bookRepository.save(book);
+        return BookDto.fromDomain(updated);
     }
 
     @Override
     public void deleteById(long id) {
+        boolean exists = bookRepository.existsById(id);
+        if (!exists) {
+            throw new NotFoundException();
+        }
         bookRepository.deleteById(id);
     }
 }
